backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
